% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCIGAR.R
\name{.parseCIGAR}
\alias{.parseCIGAR}
\title{Parse a CIGAR string into its constituent operations and counts}
\usage{
.parseCIGAR(cigar)
}
\arguments{
\item{cigar}{(Mandatory) A character string representing the CIGAR string (e.g., "5M1I3M")}
}
\value{
A data frame with two columns:
\describe{
\item{count}{A numeric vector containing the count of bases for each operation.}
\item{operation}{A character vector indicating the operation type for each segment (e.g., "M" for match, "I" for insertion).}
}
The resulting data frame helps to interpret each CIGAR operation and its associated base count for alignment purposes.
If the input is invalid or empty, the function will return an empty data frame.
}
\description{
\code{.parseCIGAR()} takes a CIGAR string and splits it into individual operations (such as match, insertions, deletions, etc.) along with their corresponding counts.
This function provides a structured breakdown of these operations for further analysis, such as visualizing alignments or processing genomic data.
}
\details{
The function uses regular expressions to split the CIGAR string into alternating numeric counts and operation types.
Invalid CIGAR strings may result in an error or unexpected behavior.
}
\examples{
# Example of parsing a CIGAR string:
BamLiner:::.parseCIGAR("5M1I3M")

}
