% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ref_seq.R
\name{.fetch_ref_seq}
\alias{.fetch_ref_seq}
\title{Fetch the reference sequence for a given read}
\usage{
.fetch_ref_seq(ref_genome, read_granges, cigar_ops)
}
\arguments{
\item{ref_genome}{A DNAStringSet or similar object representing the reference genome.}

\item{read_granges}{A GRanges object representing the read's genomic alignment.}

\item{cigar_ops}{A data frame representing the parsed CIGAR string with operations and counts.}
}
\value{
A character string representing the extracted reference sequence for the given read.
}
\description{
\code{.fetch_ref_seq()} extracts the reference sequence corresponding to the genomic location of a read. The genomic range is adjusted to account
for any soft-clipping present in the CIGAR string, which could affect the alignment at both the start and end of the read.
The function ensures that the adjusted range of the reference sequence aligns correctly with the read sequence.
}
\examples{
BamLiner:::.fetch_ref_seq(ref_genome, read_granges, cigar_ops)
}
